Create storage, sensor aggregate
Create storage module to handle abstract json based storage devices.
Added a File and a Noop based implementation.

The file based stores json into a directory and the noop does nothing
when called.

To be capable of choosing a sensor on start scan a director needs
knowledge about a sensor. The sensor are referenced by a target and need
to be registered in the director upfront.