Extend connection to use topic on OnMessage, add start method on scan
Since start.scan is required to send a start.scan message to a different
topic specialized for a specific sensor the capability to set a topic on
returning a message within a handler was needed.

On top of that the internal/test/handler now uses reflection instead of
casting so that it is more transparent when creating a new message type.