Adapt topic structure to group/aggregate/event/destination
Due to the normalization of the topic structure there is no need to
manually create a SendResponse struct but rather let the handler decide
the topic based on the MessageType.

For that a new interface
```
type Event interface {
	Event() EventType
	MessageType() MessageType
}
```
got introduced and all messages retrieved or send back must be of that
interface so that the messages/handler can create a SendResponse struct
with the correct topic.

Wit that a developer must decide which kind of message it is upfront
(cmd, info) and embedded the corresponding ID.

```
type GotScan struct {
	messages.Message
	info.EventType
	Scan
}
```

this overhead frees the handler to not have to guess to correct topic
and the developer to not having to remember the topic definitions to
manually set it.