Use cmd/example-client as smoke test
Besides being a show case for using the eulabeia structure this commit
starts using it for smoke tests as well.

When running this example client it should either stop with 0 (success)
or not for failure. This is not meant to be detailed test case but
rather if the overall functionality works as indented.