Add possibility to override configuration file via environment variables
In e.g. a container the most convenient way to override a configuration
dynamically is via environment parameter. Although it must not be the
case for all but some parameter this is a prefered way to handle
configuration.

For that this commit adds the possibility to override some fields via an
environment parameter.