Add simplified openvas package to sensor (#26)
* Add simplified openvas package to sensor

Unfortunately the currently build image for openvas is not suitable for
production use case regarding the required space (1 GB) one way to solve
that is to separate the build and run requirement.

For our sensor image this is done by building a debian package. Although
those packages are not suitable for production (e.g. we don't care about
downwards compatibility hence don't track symbols within a library) it
is fitting our use case.

This allows us to minimize the sensor image for openvas from 1 GB down
to about 255 MB.

Since the description for gvm-libs and openvas are not eulabeia relevant
it is stored underneath .docker folder.

The feed is loaded into a volume and then attached to the sensor
container.

Co-authored-by: Jaspar L. <jstach@uos.de>