Add implement preprocessor for start.scan.director
To split a huge scan message into two modify and one start message
a new preprocessor for director is written.

For testing purpose the Noop storage handler of storage is rewritten
into a memory storage.

In preparation of removal of ModifySetValue reflection method due to too
much hidden complexity director/target does not use it anymore.