Change: instead of using community feed create test.nasl
Due to the dependencies within the nasl script of the community feed we
would need to add a lot of programs to run them.

To make testing easier it is better to provide simple nasl plugins by
ourself.
test.nasl plugin sends:
- security_message(data:sec_msg);
- log_message(data:log_msg);
- error_message(data:err_msg);

with that we can verify the result handling within eulabeia.