Add: Logging framework zerolog
To be able to introduce multiple log levels zerolog is introduced.

The reason for zerolog is that it supports multiple outputs:

- Json
- CBOR
- Console

it is easy to use and has little allocations.

This will make it easy to introduce log aggregator e.g. ELK stack
without having to change logging in the modules.

The creation of the logger is done via our own package 'logging' this
will allow us to make changes in output, additional fields without
nagging the caller too much.

The different log level and our interpretation can be seen within
./docs/logging.md