Refactor: change mqtt connection to just contain an out and in channel
To be able to chain multiple publisher the mqtt connection is just
returning a in channel. The rest will be handled by a abstract handler
which contains the handler as well as publisher; this allows us to chain
multiple independent publisher without having to change the mqtt
connection each time.

WIP