Fix: sensor should not block on sensor while registering
On some circumstances it can occur that a modified message is sent in a
duplicate fashion due to multiple modify events. In this case after the
first message was processed and the register method is not running
anymore the channel got overflow and the register handler was blocking
while trying to send the notification.

This commit chanegs the sending behavior to ignore it when it cannot be
send instead of blocking it.