Merge pull request #85 from nichtsfrei/refactor-mqtt
Refactor: change mqtt connection to just contain an out and in channel

To be able to chain multiple publisher the mqtt connection is just
returning a in channel. The rest will be handled by a abstract handler
which contains the handler as well as publisher; this allows us to chain
multiple independent publisher without having to change the mqtt
connection each time.

On some circumstances it can occur that a modified message is sent in a
duplicate fashion due to multiple modify events. In this case after the
first message was processed and the register method is not running
anymore the channel got overflow and the register handler was blocking
while trying to send the notification.

This commit changes the sending behavior to ignore it when it cannot be
send instead of blocking it.

Instead of using eulabeia the logger should try to guess the name based
on the executable.

To prevent multiple starts of the same scan the example client should
verify if it is already started or not.