Refactor: add configuration possibility to mtt
Instead of hardcoding the values this commit makes it possible to
configure the mtt client. To make it more convenient to create a mqtt
instance it has the possibility to create a new client based on the
connection configuration of the toml file.

To make it easier remove unneeded handler mechanism in favor of adding
the possibility to register OnMessage for low level stuff as well as
Event handler.